external file urls can now be edited (#1)
In the initial MVP we did not implement url editing. We could not reuse the editor plugin feature because the file had a specific format that we did not want to expose to end-users through the web interface. This PR adds editing capabilities somewhere between the original create dialog for external file links and how the edit button is integrated in the editor plugin.